requirements.txt
# CTF Toolkit - Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core CLI & Output ────────────────────────────────────────────
rich>=13.7.0              # Colored terminal output, tables, progress bars
python-dotenv>=1.0.0      # .env file loading
PyYAML>=6.0.1             # config.yaml parsing

# ── Web Exploitation ─────────────────────────────────────────────
requests>=2.31.0          # HTTP client
urllib3>=2.0.7            # Underlying HTTP, disables InsecureRequestWarning

# ── Steganography ────────────────────────────────────────────────
Pillow>=10.2.0            # Image processing (LSB extraction for non-PNG)

# ── Binary Exploitation (optional but recommended) ───────────────
# pwntools>=4.12.0        # Full pwntools suite (Linux/macOS only)
#                         # Uncomment if running on Linux for full binary exploitation support
#                         # pip install pwntools

# ── Cryptography (stdlib covers most needs) ──────────────────────
# No extra deps needed — uses Python stdlib: hashlib, base64, struct, math

# ── Development & Testing ────────────────────────────────────────
pytest>=8.1.0
pytest-cov>=5.0.0
